Moved commonTools/git/ to cmake/tribits/common_tools/git/ and fixed checkin-test.py tests (broken)
I moved the commonTools/git/ directory to the directory
cmake/tribits/common_tools/git/ so that it can be reused as part of the new
TriBITS system.  I changed the name from 'commonTools' to 'common_tools' to
match the other directory names in the 'tribits' directory.

There are still several failing unit tests so this refactoring is not finished
yet.