Generalizing TriBITS call-backs to be on a Repository basis; more generalization
This is another step in the process to allow the creation of meta-builds with
the TriBITS system.  Here is what I did here:

1) Created ${PROJECT_NAME}_NATIVE_REPOSITORIES varibles that is set in the
project's top-level CMakeLists.txt file.  For Trilinos, this is just set to
"." (since Trilinos is both a Project and a Repository).

2) Created ${PROJECT_NAME}_NATIVE_REPOSITORIES that concatenates
${PROJECT_NAME}_NATIVE_REPOSITORIES and ${PROJECT_NAME}_EXTRA_REPOSITORIES

3) Broke out seprate call-back files for each of the call-back macros.

4) Pulled CTestCustom.ctest.in out of tribits and into a project-specific
location since it should be customized on a project-speicfic level.  I also
tried to remove some of the duplicated code.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Disabled Packages: PyTrilinos,TriKota,Optika,Sundance,Stokhos
Enabled all Packages
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => FAILED: passed=1284,notpassed=1 => Not ready to push! (10.12 min)
3) SERIAL_RELEASE_SS => passed: passed=1171,notpassed=0 (5.63 min)
WARNING: Forced the push!