Changing from CTEST_BUILD_NAME to TDD_BUILD_NAME in env
The name CTEST_BUILD_NAME is hard-coded into CTest and is aready read off the
env by the script TribitsCTestDriverCore.cmake.  Therefore, you can't use this
name for the TDD system otherwise it will override the name for the inner
builds (which is bad, trust is, we did that).  This varible does not seem to
be set in any of the working automated build scripts so this should not break
backward compatibility.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Disabled Packages: PyTrilinos,TriKota,Optika,Sundance,Stokhos
0) MPI_DEBUG => passed: passed=92,notpassed=0 (1.35 min)
1) SERIAL_RELEASE => passed: passed=92,notpassed=0 (0.93 min)