Fixing 'make dashboard', Cleaning up selection of notes files, update doc, fix test dirs grep
When I tried to run the 'make dashboard' target, it did not work for Trilinos.
I don't know how this happened but the experimental_build_test.cmake script is
currently not unit tested so I can see why it could break without people
knowing it.  I had to pass through PROJECT_SOURCE_DIR.

I aggregated the selection of the notes files into more logical places, just
before the notes are submitted to CDash in each case.  This helps to make
clear the ordering of the notes files.  However, since the notes files are
submitted in two calls, they have to be defined in different places.

I also updated the documentation in the experimental_build_test.cmake script.

The last thing I did was to fix many of the TriBITS unit tests that were
broken when a symbloc directory name was used.  In one part of CMake, the
symbolic link was used and in another part, the true path was used.  I think
this was due to the use of the 'make absolute' CMake command.  This was
messing up a unit test.  Anyway, I changed the grep of the unit test to be
pretty good (but not perfect).

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Disabled Packages: PyTrilinos,TriKota,Optika,Sundance,Stokhos
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=92,notpassed=0 (1.30 min)
1) SERIAL_RELEASE => passed: passed=92,notpassed=0 (0.93 min)