Major refactor to coordinate output, turn off colors, show debug info, namespace egdist arguments
With this refactoring, when you run with egdist --no-pager, the output from
egdist and the inner eg commands now all line up correctly because we are now
using the Python subprocess module.

To make the output cleaner to read when piped to a file, I also added the
option --dist-no-color.

All of the options pulled out for egdist are now prefixed by "--dist-" to
disambiguate them from raw eg/git arguments.  This also affects the
--extra-repos option which is now --dist-extra-repos.