Adding assert and test for packages and subpackages with self references
Believe it or not, the TriBITS system prior to this commit would allow a
subpackage to reference itself.  Unfortunately, an unnamed TriBITS project
developer used this ability to put in several subpackage self references that
(silently) broke the generation of the CDashSubprojectDependencies.xml file
due to an infinite recursion.

Actually, the prior behavior would not allow a package to reference itself but
it produced a less clear error message.With the new behavior, a better error
message is produced in the case of package self-reference.

As of this commit, there are unit tests that test the behavior of TriBITS when
a package or a subpackage tries to reference itself as a dependency.

Now I must fix that other project or this will break their TriBITS build!

Build/Test Cases Summary
Enabled Packages: Teuchos, Denovo, VRIPSS
Disabled Packages: PyTrilinos,Optika,Sundance,Stokhos
0) MPI_DEBUG => passed: passed=95,notpassed=0 (3.08 min)
1) SERIAL_RELEASE => passed: passed=93,notpassed=0 (2.90 min)
2) MPI_DEBUG_SS => passed: passed=597,notpassed=0 (22.65 min)
3) SERIAL_RELEASE_SS => passed: passed=338,notpassed=0 (9.29 min)
4) MPI_DEBUG_SS_STATIC => passed: passed=597,notpassed=0 (26.19 min)
5) SERIAL_RELEASE_SS_STATIC => passed: passed=338,notpassed=0 (8.97 min)
Other local commits for this build/test group: d5c8aab