Setting  ${REPOSITORY_NAME}_SOURCE_DIR in TRIBITS_READ_PACKAGES_PROCESS_DEPENDENCIES_WRITE_XML()
When run from the TribitsDumpDepsXmlScript.cmake cmake script from the
checkin-test.py script, the variables ${REPOSITORY_NAME}_SOURCE_DIR were not
getting set.  This is not such a big deal except that current Exnilhilo's
Insilico package's Dependencies.cmake currently uses logic based on if
Scale_SOURCE_DIR is set or not.  Therefore, even when the Scale TriBITS repo
is provided it was not added the Insilico subpackage correctly.

This is a pretty obscure thing.  I am not sure that it is worth writing a unit
test for this because longer term better TriBITS dependency logic will make it
unnecessary to use such logic in Dependenices.cmake files.

Even through there is not a unit test, I changed other code inside of the
cmake macro that uses the variables ${REPOSITORY_NAME}_SOURCE_DIR so if it
does not get set for some reason, the code will break and the unit tests will
fail.  That is good enough for me considering everything.

Build/Test Cases Summary
Enabled Packages: Teuchos, CTeuchos
Disabled Packages: CTrilinos,ForTrilinos,PyTrilinos,Didasko,Mesquite,Phdmesh,Pliris,Claps
0) MPI_DEBUG => passed: passed=101,notpassed=0 (1.62 min)
1) SERIAL_RELEASE => passed: passed=99,notpassed=0 (1.19 min)
2) MPI_DEBUG_SS => passed: passed=104,notpassed=0 (4.72 min)
3) SERIAL_RELEASE_SS => passed: passed=100,notpassed=0 (2.57 min)