TrilinosFramework: Fix performance issue with export files
bug 5365

The creation of the export files was taking inordinately long due to having to
build the list of packages that a package depended on. The implementation of
this tried to build that information up after the fact and was very slow as it
had to walk the complete dependency tree for pretty much every package to
determine the indirect dependencies.

The logic was changed to instead build a full list of the packages that are
depended upon when a package's dependencies are being determined. This makes it
possible to quickly determine the indirect dependencies since they are now
stored for each package. This system has obtained a roughly 20x speed up of the
configure. The variable <project_name>_ENABLE_EXPORT_MAKEFILES was left to
support backwards compatibility, however, configure times likely won't be
affected too much by it now as the largest cost now should be writing the export
files.

There is still likely some improvement that can be gained in the export system
but nothing as dramatic as what was fixed and likely at the cost of significant
developer time. If the package list grows significantly, however, a sort that is
being used to ensure the package list is in dependency order may become
unscalable.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Enabled all Packages
0) MPI_DEBUG => passed: passed=1422,notpassed=0 (117.58 min)
1) SERIAL_RELEASE => passed: passed=1421,notpassed=0 (53.67 min)