Fix issue with pthread and the makefile export system.
The change to pthread highlighted a known deficeincy in our tpl system w.r.t.
the export system. The change made it possible to have a linker option be the
intended input to the target_link_libraries call. That is something that cmake
supports so we shouldn't disallow that. However, the way we pass the option to
that call is through TPL_Pthread_LIBRARIES, which was intended to store only
library names or paths. The makefile export system relied on this behavior and
would add the necessary link options to anything that was not a path. This is
something that has caused issues before, specifically with MKL and the options
Intel suggest using when linking MKL. Since the use case of the changes to
pthread is reasonable, as is the use case for MKL it doesn't make sense to have
the makefile export system hold those usages back. The general tribits build
system was fine with non-libraries being in the TPL_<tpl name>_LIBRARIES
variables since it was generally just passed through to cmake so this doesn't
affect the general build.

Also added better protection around the usage of find_package(Threads).
Initially any threading library that was found was assumed to be pthreads,
however Threads can find other threading libraries. It has also been protected
such that if a user gives a path to find pthread we assume that they want us to
use that instance. The Threads library doesn't provide a way to look in a
specific location for pthread, instead relying solely on if it can successfully
build/link an executable. In that case the call to Threads is skipped entirely
and we use the tribits tpl system to ensure that what the user specifies is
used.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Enabled all Packages
0) MPI_DEBUG => passed: passed=1449,notpassed=0 (118.05 min)
1) SERIAL_RELEASE => passed: passed=1444,notpassed=0 (50.88 min)