Updated the FindPythonInterp.cmake file
The previous version of FindPythonInterp.cmake ran python -V and took
the last output argument (-1) as the version number.  It turns out
this is not a valid assumption.  I was running with an
anaconda-installed version of python, for which python -V returns

  Python 2.7.3 :: Anaconda 1.4.0 (x86_64)

giving "(x86_64)" as the erroneous version number.  (Anaconda is a
free python distribution aimed at big data.)  I changed
FindPythonInterp.cmake to call

  ${PYTHON_EXECUTABLE} -c "import sys; print sys.version.split()[0]"

which should always give the proper version number, even if we are
running a specialized version of python.