Allow creation of flexible export makefiles to support wrapping external projects (Trilinos #5717)
The primary driver for this refactoring is to support more flexible generation
of export makefiles to aid in wrapping external projects as Trilinos packages
that depend on Trilinos and already use the Trilinos export makefile.  For
more details, see Trilinos #5717.

The primary changes in this commit are the change for the old
TribitsPackageWritePackageConfig.cmake file to the new
TribitsWriteClientExportFiles.cmake file.  This creates a new function:

  TRIBITS_WRITE_FLEXIBLE_PACKAGE_CLIENT_EXPORT_FILES(
    PACKAGE_NAME <pakageName>
    [EXPORT_FILE_VAR_PREFIX <exportFileVarPrefix>]
    [WRITE_CMAKE_CONFIG_FILE <cmakeConfigFileFullPath>]
    [WRITE_EXPORT_MAKLEFILE <exportMakefileFileFullPath>]
    [WRITE_INSTALL_CMAKE_CONFIG_FILE]
    [WRITE_INSTALL_EXPORT_MAKLEFILE]
    )

That function is pretty much the entire purpose of this commit.  It was no
small task.

Other than the refactoring to create the above function, there were minor
changes in other files to support this effort.  The most important changes to
be careful about are those that affect the basic function of compiling,
linking, and running software.

The only files that are really involved in setting up, modifying, and using
the critical intra-package dependancy variables to build and link software
are:

    TribitsPackageMacros.cmake
    TribitsLibraryMacros.cmake
    TribitsAddExecutable.cmake

The changes I made to the other files are pretty minor for the most part.  The
major changes I made were to the file TribitsPackageWritePackageConfig.cmake
which is now the better named file TribitsWriteClientExportFiles.cmake.

I looked carefully at the diffs for these and other files to decide how
extensive of testing I will need to do before pushing.

Looking over all of these files, the only real changes I made where:

(*) Added variable ${PROJECT_NAME}_GENERATE_EXPORT_FILE_DEPENDENCIES to
determine whether ${PACKAGE_NAME}_FULL_EXPORT_DEP_PACKAGES gets generated or
not.  This gets set to ON if ${PROJECT_NAME}_ENABLE_EXPORT_MAKEFILES==ON or
${PROJECT_NAME}_ENABLE_INSTALL_CMAKE_CONFIG_FILES==ON.  Also, individual
packages can set ${PROJECT_NAME}_GENERATE_EXPORT_FILE_DEPENDENCIES=ON
independent of generating the default export files if they need to guarantee
that these dependencies will be generated no mater what.

(*) Modified the system so that you can set
${PROJECT_NAME}_ENABLE_EXPORT_MAKEFILES=ON with
${PROJECT_NAME}_ENABLE_INSTALL_CMAKE_CONFIG_FILES=OFF and still generate the
export makefiles.  Roger blew a lot of time a few weeks ago trying to figure
that one out.  That just makes common sense does it not?

(*) The file TribitsPackageConfigTemplate.export.in was augmented so that when
used to generate a package-level Makefile.export.<posfix> file, it can be used
as a drop-in replacement for the project-level Makefile.export.<projectname>
file.  This is needed for the CASL_MOOSE use case, for example, and it just
makes sense.  Backward compatibility of function has been maintained
(i.e. same variables with the same content) except duplicate lib dirs and
include dirs have been removed.

(*) Added lots of if-protected debug-mode printing.  I have taken to the
convention of protecting debug print statements with the name of the cmake
function with _DEBUG_DUMP after it.  That allows me to only turn on what I
want to see and not get a sea of output that I can't make heads or tails of.

(*) Added some more documentation.

(*) Name changes for writing export file functions/files:

- The old name TribitsPackageWritePackageConfig.cmake was fine when this file
contained only the macro to write package ${PACKAGE_NAME}Config.cmake files
but it actually now contains macros that write export makefiles and the
project-level ${PROJECT_NAME}Config.cmake and export makefile.  The new name
TribitsWriteClientExportFiles.cmake is more correct and gives the intent of
the file.

- Changed TRIBITS_WRITE_PACKAGE_CONFIG_FILE() to
TRIBITS_WRITE_PACKAGE_CLIENT_EXPORT_FILES(): The name
TRIBITS_WRITE_PACKAGE_CONFIG_FILE() was fine when this function only wrote
${PACKAGE_NAME}Config.cmake files but it totally leaves out the function of
writing export makefiles.  The new name
TRIBITS_WRITE_PACKAGE_CLIENT_EXPORT_FILES() is more correct in that it writes
multiple "FILES" and not a single "FILE" and the name "CLIENT_EXPORT_FILES" is
a generic term that can apply to both ${PACKAGE_NAME}Config.cmake and export
makefiles.

- Changed TRIBITS_WRITE_PROJECT_CONFIG_FILE() to
TRIBITS_WRITE_PROJECT_CLIENT_EXPORT_FILES(): The reason for this name change
is the same as above only for the project files instead of the package-level
files.

Details on how this was tested are in Trilinos #5917.