Removing wrong and evil INCLUDE_DIRECTORY() and LINK_LIBRARIES() commands (Trilinos #5922)
The FindTPLPETSC.cmake module what was added included the commands:

  INCLUDE_DIRECTORIES(${petsc_include})
  LINK_LIBRARIES(${petsc_lib})

This is super evil and violates the very functioning of TriBITS TPL
dependencies.

I found this because we were getting failing tests with Hydra-TH in VERA where
it was pulling in the include and libs for this version of "PETSC" which
conflicts with a specialized version of PETSC (which includes an old static
version of ML).

There may be other defects with the FindTPLPETSC.cmake module pulled in from
MPACT.  If I see any other major issues, I am going to gut the thing and go
back to a simpler standard form that was working fine.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework, LIME
Disabled Packages: CTrilinos,ForTrilinos,PyTrilinos,Didasko,Mesquite,Phdmesh,Pliris,Claps
Enabled all Forward Packages
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=122,notpassed=0 (21.18 min)