Making enabling a package the same as enabling all its subpackages (Trilinos #5965)
Previously, enabling a package was *not* identical to enabling it subpackages
when enabling forward packages for testing downstream changes.  If a
downstream SE package only depended on a subpackage <Package>A but not
<Package>, then that downstream package would not get enabled.  This was very
bad for testing purposes and is very confusing for users tha seem to expect
this behavior (I did).  Now enabling a package is identical to enabling all of
its required and optional subpackages (if optional subpackages are to be
enabled or not).

The driver for this change is usage in CASL VERA (see Trilinos #5965 and VRI
Kanban #3098).

This technically changes backward compatibility but only in cases where the
user enabled foward dependencies which really only developers do for testing
purposes.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework, Teuchos
Disabled Packages: CTrilinos,ForTrilinos,PyTrilinos,Didasko,Mesquite,Phdmesh,Pliris,Claps
0) MPI_DEBUG => passed: passed=224,notpassed=0 (6.32 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)