Adding install target for <Project>RepoVersion.txt from tarball source
When building a tarball, the <Project>RepoVersion.txt file gets put into the
tarball.  This new install hook looks for that file in the source tree and if
it finds it, installs it into the installation tree.

This is an important part of version tractability for releases.

I tested this manually and it worked as expected.