Breaking up packaging callback into repo and project versions, specialize soruce gen types (Trilinos #5970)
Now there are two callback files, one for the repo and one for the project.
The file CallbackDefineRepositoryPackaging.cmake get called always (if the
repo isincluded at all) and the file CallbackDefineProjectPackaging.cmake only
gets called if this is the base TriBITS project.  No more if statement looking
at the project name.

Added the cache variable ${PROJECT_NAME}_CPACK_SOURCE_GENERATOR to set
CPACK_SOURCE_GENERATOR and allow the project to define the default
${PROJECT_NAME}_CPACK_SOURCE_GENERATOR_DEFAULT.  You can't ask for any more
flexibility or specialization than this!