Shortened the history paragraph on TriBITS, added PDF to version tracking
Now that the longer (but still interesting) history of TriBITS is in the new
TribitsDevelopersGuide.rst document, I cut down the history in the intro to
the bare minimum.  The point is to make it clear that TriBITS was motivated by
need and built on the successful Trilinos package model that was also
successful in CASL VERA.

The PDF file should be under version control too so that it is easier to
publish.