Fixed repo source dir to not end in '.' for default project native repos
I have finally updated TriBITS code and I should have the tests in place to
show that ${REPO_NAME}_SOURCE_DIR will *not* end in "/." for the default
native repo.  Having this in the directory name was ugly and it messed up file
regexes (like in cpack excludes).

This commit also adds testing of the tarball creation process.