Allow setting ${PROJECT_NAME}_ENABLE_SECONDARY_STABLE_CODE and ${PROJECT_NAME}_ENABLE_SECONDARY_TESTED_CODE at the same time
I changed this becuase the VERA CI testing is setting both for some reason and
it is failing.  It is little dangerous to allow both of these to be set but in
the short term this maintains better backward compatibility.

Build/Test Cases Summary
Enabled Packages: TriBITS, CTeuchos
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=136,notpassed=0 (0.98 min)
1) SERIAL_RELEASE => passed: passed=131,notpassed=0 (0.91 min)