Adding RST doc for TRIBITS_CTEST_DRIVER() and shorten column length
Here I have adeed the standard header for the funciton TRIBITS_CTEST_DRIVER()
that will go into the developers guide.  This will be a place to document the
behaivor of this intead of refering to the TribitsCTestDriverCore.cmake script
(which is never directly run by itself as a ctest -S script).