Edited up through 'References', added section on reg. backward compat.
I edited all of the remaining sections up through the 'References' section.  I
also took the contents of the file README.DEPRECATED_CODE and put it under the
'Regulated Backward Compatibility' section and formated for RST and linked in
with the rest of the document.  We are now at 159 PDF pages!

I removed the file README.DEPRECATED_CODE because we don't need it anymore.