Adding documentation, shorten cache var name some, clean up print output (Trilinos #6136)
This should be the final commit for the auto detection of C++11 compiler
options.  Here is what I did:

1) Change name of ${PROJECT_NAME}_TRIBITS_CXX11_FLAGS to
${PROJECT_NAME}_CXX11_FLAGS and made it an "advanced" cache var.  The
convention with TriBITS is not to put "TRIBITS" in the name of most variables.
Based on that, every variable that TriBITS defines would need to have
"TRIBITS" in the name.  Prefixing by the project name should be enough.

2) Change from printing "CXX11" to "C++11" to make it more understanable and
easier to grep in the output.  For variables we should use CXX11 but for
regular output, C++11 looks better.

3) Updated section on C++11 support to mention <Project>_CXX11_FLAGS and
automatic compiler flag detection.