[TriBits] remove duplicates from header and sources lists
Many Trilinos libraries include certain header and source files more than
once. As a consequence, some files are installed more than once. While this
is not a huge problem per se, it hides the overrides that are a problem.
See bug #6186.

This commit just removes the duplicate entries of the HEADERS and SOURCES
lists for TRIBITS_ADD_LIBRARY(). The duplicate installation count goes down
from 62 to 30 with this.