Updated sections on building single object files
Newer versions of CMake seem to be naming the object file target for source
files SomeSourceFile.cpp as SomeSourceFile.o.  And the object file is stored
in a subdirectory.  These updated instructions should work for any version of
CMake that stores the object files under the build directory of its Makefile
that creates it.