Test TRIBITS_DETERMINE_IF_CURRENT_PACKAGE_NEEDS_REBUILT() with WrapExternal
I have updated this test to copy TribitsExampleProject to the test dir and to
do several several reconfigure and rebuilds testing out rebuilds with changes
upstream files in source and build dir.

This software is not unit tested right now so that is the only test we have.
However, it is a pretty nice acceptance test and it does not take that much
longer to run.