Only export library targets that are really cmake libraries (Trilinos #6168)
The IMPORTEDLIBS <lib0> <lib1> ... argument currently addes the <lib0> <lib1>
etc. to the package's ${PACAKGE_NAME}_LIBRARIES variable.  This is bad because
this list of libraries was being passed to EXPORT(TARGETS ...) and was causing
a configure-time failure when "IMPORTEDLIBS m" was used (see Trilinos #6068,
comment 37).

Really we should *not* be adding IMPORTEDLIBS to ${PACAKGE_NAME}_LIBRARIES.
They are already being linked into CMake-built libraries that CMake knows
about.  However, it seems that the Trilinos package TriKota is currently
pulling in Dakota-built libraries using IMPORTEDLIBS so I think we may break
the export files if we remove IMPORTEDLIBS from ${PACAKGE_NAME}_LIBRARIES.

The temporary solution I came up with was to filter the set fo libraries in
${PACAKGE_NAME}_LIBRARIES with IF (TARGET <libName>) and only pass those
libraries to EXPORT(TARGETS ...).  That seems to do the trick.  Now we can
pass in "IMPORTEDLIBS m" and "m" will not be passed to EXPORT(TARGETS ...) but
any IMPORTEDLIBS that are actually CMake targets will get passed to
EXPORT(TARGETS ...).  In manually verified this with the Trilinos package
Zoltan.

There is currently no automated testing for any of this export file stuff.
That is a situation that needs to change pretty darn soon.