Fixing TRIBITS_ADD_ADVANCED_TEST() handling of STDERR with OUTPUT_FILE (Trilinos #6181)
It turns out that EXECUTE_PROCESS() does not allow OUTPUT_FILE and ERROR_FILE
to point to the same files.  On Unix and Linux, stderr overwrites stdout all
together.  On MS Windows, you get an error if both point to the same file.
See some of the later comments in Trilinos bug #6181 that describe the
problem.

The way that I fixed this problem is to actually just always use
OUTPUT_VARIABLE and ERROR_VARIABLE to capture console output into the same
variable and then write the file specified by the OUTPUT_FILE argument
afterward.  This captures both STDOUT and STDERR.  However, it does not
capture them in the correct order.  However, contrary to documentation
provided by CMake, STDOUT and STDERR may not be interleaved correctly,
especially if MPI is used.  In fact, it may not even output in the correct
order in serial if the machine is loaded.  See the test:

    TriBITS_CTestScripts_cmnd_3_args_1_overall_working_directory

for more details.

I also also put in a warning in the documentation for
TRIBITS_ADD_ADVANCED_TEST() about the usage of OUTPUT_FILE and the strange
behavior of STDOUT and STDERR.

Build/Test Cases Summary
Enabled Packages: TriBITS
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=135,notpassed=0 (0.56 min)
1) SERIAL_RELEASE => passed: passed=135,notpassed=0 (0.53 min)
Other local commits for this build/test group: a6988b3