Converting documentation index to simple index.html file
It is silly to use DocUtils to generate a simple index.thml file.  Also,
DocUtils does not allow you to create multiple <a href="<link1>" PDF</a>, <a
href="<link2>" PDF</a>, etc. in the document.  This is stupid.

Just get rid of the fluff and have a simple index.thml file.