Adding the origin dir repo name and SHA1 to commit summary, listing tracking branch
With this commit, I added the name of the origin repo dir name and the origin
repo SHA1 to the commit message summary.  Now this looks like:

    Automatic snapshot commit from TriBITS at c8d7a01

instead of just

    Automatic snapshot commit

I also updated the script to find out the remote tracking branch and dislay
that instead of assuming origin/master.