Fixing nasty bug where TriBITS checkin-test.py config used instead of project's (TriBITS #5)
As explained in the TriBITS Github issue #5, The TriBITS checkin-test.py
configuration was always getting used instead of the proejct's configuration
file.  This was caused by the commit:

  5c2aa1d "Allow for usage of the checkin-test.py script for TriBITS itself."
  Author: Roscoe A. Bartlett <bartlettra@ornl.gov>
  Date:   Fri Sep 5 15:58:24 2014 -0400 (4 days ago)

  A       project-checkin-test-config.py

which exposed a latent defect added way back when TriBITS was first split off
from Trilinos. The problem was that the base directory for the found
<projectDir>/project-checkin-test-config.py was append *not* prepended to
sys.path before including the module with:

    __import__(moduleName).configuration

This was very bad becuase the TriBITS base path automatically gets *prepended*
by python into sys.path just by the checkin-test.py script running the
checkin-test.py script.  Therefore, when I added the file:

   TriBITS/project-checkin-test-config.py

which got snapshotted to:

   Trilinos/cmake/tibits/project-checkin-test-config.py

this logic resulted in this file getting selected for the project's
checkin-test.py configuration file, not the project's:

   Trilinos/project-checkin-test-config.py

file.

By picking up the TriBITSProj project's project-checkin-test-config.py file,
it got --enable-packages=TriBITS and also ignore all of the other
Trilinos-specific MPI_DEBUG and SERIAL_RELEASE options.  Very bad.

Also, as part of this commit, I figured out how to set the default for
--src-dir to the same directory where the project's
project-checkin-test-config.py is found.  That means that checkin-test.py will
automatically pick up the projects project-checkin-test-config.py file as long
as the checkin-test.py script is symlinked into the project's base directory
tree, no matter where it get TriBITS from.  This also allows the
checkin-test.py script under the TriBITSProj build to automatically find the
right project directory.

This is really a great commit if you take this all in!

For GitHub: Fixes #5