Fixing usage of NativeRepositoriesList.cmake
When a TriBITS repo's PackagesList.cmake file uses the macro
TRIBITS_REPOSITORY_DEFINE_PACKAGES() and the repo was defined in the
NativeRepositoriesList.cmake file, the configure would fail because of an
error in logic that would not propertly set the REPOSITORY_NAME variable
before reading in the repo's PackagesList.cmake and TPLsList.cmake files.
This logic has now been fixed.

Here I also added an auotmated test for the usage of the
NativeRepositoriesList.cmake file.  I put in a hook to allow this file to be
specified using ${PROJECT_NAME}_NATIVE_REPO_FILE_OVERRRIDE for unit testing
purposes.  This variable is undocumented on purpose.  I don't want users to
know about it or use it.

Really I want NativeRepositoriesList.cmake to go away and for projects to use
ExtraRepositoriesList.cmake but first I need to implement support for
pre-extra repositories.