Changing selection of CTest based on exact path (Trilinos #6189)
With the new CMake version 2.8.11, the file name 'ctest' is found not only in
bin/ctest but also in share/2.8/ctest so the glob logic in
LocateCTestAndRunScript.cmake failed.  I think this logic was there for back
in the days when download_cmake.py did not cleanly install cmake/ctest under
<install-dir>/bin.  This caused the inner build to fail since it could not
find ctest.

I have changed this to just pick out the exact ctest path.  This seems to work
and is a lot safer.