Use REALPATH wherever possible
The relocatabilty of the Trilinos installation demands all paths in the
export files, specfically the include paths, to be relative, e.g.,

       "${CMAKE_CURRENT_LIST_DIR}/../../../include/"

This expands to a different path for every package, so for the entire
project (e.g., Trilinos) the consumer compile lines will include a separate
`-isystem` include path for *every* package, even if they all point to the
same absolute folder (which is often the case).

This commit applies REALPATH in the export files such that the paths are
expanded just before they are used. This allows CMake to rule out duplicates.
Compilation lines are now generally much shorter and easier to debug.