Addressing several TRIBITS_CTEST_DRIVER() defects (PHI Kanban #3436, #3231, #3086, #3470)
PHI Kanban #3436: Making ${PROJECT_NAME}_EXCLUDE_PACKAGES disable pakages even
if set by ${PROJECT_NAME}_PACKAGES.  However, to maintain backward
compatiblity, I had to allow packages listed in ${PROJECT_NAME}_PACKAGES to
override regular disables. This allows you test EX pakages that are hard
disabled by default by listing them in ${PROJECT_NAME}_PACKAGES.  However, we
need for ${PROJECT_NAME}_EXCLUDE_PACKAGES to override and disable packages set
in ${PROJECT_NAME}_PACKAGES.  I added a good unit test for this and updated
other unit tests to look for the right things in the output.

PHI Kanban #3231: I moved the delete of the CMakeCache.txt file to after we
decide if we are going to be configuring any packages.  This is critical since
the build directory will not be deleted if a CMakeCache.txt file is not
present.  I added comments to discuss this issue in the *.cmake file.  This
should address defect in the CI server not starting with a fresh build dir in
PHI Kanban #3231.

PHI Kanban #3086: I added the delete of the CMakeFiles/ directrory along with
the CMakeCache.txt file.  For later versions of CMake, they must be deleted
together or a configuire from scratch will not work.  I commented on that in
the *.cmake file.

PHI Kanban #3470: I added the options
CTEST_GENERATE_OUTER_DEPS_XML_OUTPUT_FILE and
CTEST_SUBMIT_CDASH_SUBPROJECTS_DEPS_FILE to allow disabling the generation of
the <Project>PackageDependneices.xml file and the
CDashSubprojectDependencies.xml file, respectively.  This is to speed up
manual testing on large projects where it is expensive to generate these files
(e.g. VERA) but the option CTEST_SUBMIT_CDASH_SUBPROJECTS_DEPS_FILE will be
used to avoid the generation of CDashSubprojectDependencies.xml for CI builds
which is currently taking a long time for VERA.

Build/Test Cases Summary
Enabled Packages: TriBITS
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=138,notpassed=0 (0.72 min)
1) SERIAL_RELEASE => passed: passed=138,notpassed=0 (0.60 min)