Implementing new Primary Meta-Project Packages (PMPP) selection (TriBITS #9)
This defines new variables:

  ${REPOSITORY_NAME}_REPOSITORY_CONTAINS_NO_PRIMARY_META_PROJECT_PACKAGES
  ${REPOSITORY_NAME}_REPOSITORY_CONTAINS_NO_PRIMARY_META_PROJECT_PACKAGES_EXCEPT

and removes the variables:

  ${REPOSITORY_NAME}_NO_IMPLICIT_PACKAGE_ENABLE
  ${REPOSITORY_NAME}_NO_IMPLICIT_PACKAGE_ENABLE_ENABLE

(which were a hack).

Also removed were the functions:

   TRIBITS_REPOSITORY_IMPLICIT_PACKAGE_ENABLE_IS_ALLOWED()
   TRIBITS_APPLY_REPOSITORY_NO_IMPLICIT_PACKAGE_ENABLE_DISABLE()

The updated behavior is deined in TriBITS Github issue #9.

This will also be fully documented in the TribitsDevelopersGuide.* document in
the next commit.

I added the macro SELECT_FINAL_SET_OF_PACKAGES_TO_PROCESS() that modifies the
set of packages that are exlicitly processed in the loop.  This makes for
nicer output and easier debugging becuase you can see exactly the packages
that will be directly processed before the loop starts and every iteration of
the loop with process a pacakge and not just skip it.

Some of the existing tests had to be modified because some strings are no
longer printed.  However, the tests should be just a strong because the look
at the final list of packages that CTest/CDash will process.

Build/Test Cases Summary
Enabled Packages: TriBITS
Enabled all Packages
0) MPI_DEBUG => passed: passed=641,notpassed=0 (36.70 min)
1) SERIAL_RELEASE => passed: passed=548,notpassed=0 (15.36 min)
Other local commits for this build/test group: 5ac058d, cb72771