Removing print of packages enables in summary email header
This commit removes the printing of the packages in the summary email
header. That list was not even correct because you can run differnet build
cases like the PT --default-builds and other ST --st-extra-builds and even EX
--extra-builds and all of these build could have different sets of enabled
packages.  You can also run the individual builds separate using different
values for --enable-all-packaes and --enable-packages.

The added benefit of removing this is that we don't have to call
getEnablesLists() a second time which is super expensive if there are a lot of
changed files and you have a lot of packages.