Making --enable-all-packages=on skip package change check
Before, if you set --enable-all-packages=on but did not set
--enable-packages=<somepkg>, it would still look for changed packages.  This
makes no sense.  If you are enabling all packages you should not bother to
look for changed packages.

Making this change has two important consequences.  First, for a meta-project
that has defined primary meta-project packages (PMPPs), then you want the user
to be able to test all of those packages just by setting
--enable-all-package=on and not have the checkin-test.py scrpt look for
changed upstream packages to enable that are not relevant to the meta-project.
A second very important consequence is that when you have tens of thousands of
changed files with a project with hundreds of packages, the package change
detection code can run very very slow currently.  Therefore, there needs to be
a simple way to turn off that change detection and just test everything the
(meta) project cares about by just setting --enable-all-packages=on.