Revert "Removing print of packages enables in summary email header"
This reverts commit 4f497cc2f1ef88b62b00eb0908f9bcfba4d8003c.

I reverted this because this also removed the enables info from the amended
commit message.  If the DID PUSH email is not sent out to a general list, this
amended commit message is the only info we have for what was tested.

This issue needs to be addressed differently.  I will write a TriBITS ticket
defining what we need at some point.