Revert "Removing print of packages enables in summary email header"
This reverts commit 9c17d1f29bd78406e5424534835d337b6c18fb82.

I reverted this because this also removed the enables info from the amended
commit message.  If the DID PUSH email is not sent out to a general list, this
amended commit message is the only info we have for what was tested.

This issue needs to be addressed differently.  I will write a TriBITS ticket
defining what we need at some point.