Making CMake 2.8.11 requried (Trilinos #6189)
This makes the minimum required CMake version 2.8.11.  This will allow TriBITS
to use some newer features and simplify the build system.

As part of this upgrade, I was able to remove some if statements based on
CMake version.

However, this commit does *not* try to clean up any of the code that uses
GET_FILE_COMPONENT() to create CMAKE_CURRENT_LIST_DIR.  It seems that when
CTest runs that CMAKE_CURRENT_LIST_DIR is *not* set to the current file's base
directory.  Instead, when ctest is running, CMAKE_CURRENT_LIST_DIR gets set to
directory where ctest is run out of.

We really need some automated testing for the TDD system in order to help
catch issues like this.

Build/Test Cases Summary
Enabled Packages: TriBITS, CTeuchos
0) MPI_DEBUG => passed: passed=143,notpassed=0 (0.97 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
Other local commits for this build/test group: 8269474