Switch sleep and touch and increase sleep to 1 full sec
It seems there are systems like trilinos-test2 where file time stamps returned
from:

      find . -type f -printf '%T@ %p\n'

are only accurate to 1 second.  Therefore, to show a later time stamp, you
have to wait a full sec and then touch a file so that that above command shows
a later time stamp.  Having to wait another sec is unfortunate but this is a
important test to determine if the trick implemented in
TRIBITS_FIND_MOST_RECENT_FILE_TIMESTAMP() will work on that system.

Hopefully this will resolve all of the remaining TriBTS test failures when run
by Trilinos.