Replacing a O(n) find with a O(1) lookup (PHI Kanban #2719)
This is the start of trying to reduce the runtime for configures.  Here, I
moved up setting ${PACKAGE_NAME}_SORUCE_DIR as soon as we know if the SE
package ${PACKAGE_NAME} is defined and its source dir exists or not.  Now we
can use O(1) IF (${PACKAGE_NAME}_SORUCE_DIR) in logic instead of O(n) FIND()
operations.