Fixing include guards in export XXXConfig.cmake files (TriBITS #11)
Before, the include guard `_INCLUDE` was getting used when used in flexible
mode (not what you want) and the name of the last package was getting used in
the <ProjectName>Config.cmake file.  This would have caused it to conflict
with the include guard from the package's include guard.  For example, with
TribitsExampleProject (TriBITSExProj), TriBITSExProjConfig.cmake was using the
include guard PackageWithSubpackages_INCLUDED.  That is not good.

With this commit, this is fixed.

I will not try to add some automated tests for the generated export files.