Generating export files (but not testing them yet) (TriBITS #11)
Here I just turned on the generation of the export files but I am not testing
them yet.  At least this way I can see what these files look like and we can
be assured that we are not crashing the configure.  That alone would have
stopped one of my previous bad pushes.