Added a validation test for genrated <Package>Config.cmake (TriBITS #11)
Here I have updated the test TribitsExampleProject_ALL_PT_NoFortran so that it
installs the project into an install directory then it reads in one of the
<Package>Config.cmake files from the install dir to make sure that it is
setting the right variables.

It was a bit of a pain to set up this test but test represents the very first
automated test for generating, installing, and reading in
<Package>Config.cmake files.

I have not added a test for the project-level <Project>Config.cmake file.  But
we can just copy and paste the files to create a test of the project-level
file.  There is also no validate test for the Makefile.export.<Package> files
but that test would be created in a similar way.