Adding -VV to all inner ctest calls
When a test fails on another platform, I can't see why it failed.  There is
currently a failing inner test for the TribitsExmapleProject that I need to
track down (e.e. the RayTracing test).  By using -VV we should be able to see
the full test detail.  These are not big tests so this should not add much to
the output that CDash has to handle.