Improved performance and outputting (PHI Kanban #2719)
Before, the overall most recent modified file was not getting printed
correctly.  The correct file with the correct timestamp was being returned but
the output was confusing.

As part of this I also improved the outputting to show the relative base
directories better, now relative to the project base dir.

I also improved performance by not running "ls" on the files in the
intermedate directories unless we are printing them out and I changed to only
look for modified files in the parent packages, not all of the subpackages.

It looks like for the CASL_MOOSE package, these changes dropped the time from
24s to 16s.