Improving accuracy to 0.001 sec (PHI Kanban #2719)
It turns out that the date command will return nanoseconds as well.  The hard
part is that CMake MATH() does not support decimal numbers.  It is amazing how
hard it is to subtract decimal numbers when all you can do is subtract
integers.

I really need to gather up this and the multiplication code I wrote for
TIMEOUT scaling and create DecimalNumberMath.cmake module.  Then I will never
have to write this type of stupid code again.