Added <Project>_ENABLE_PACKAGE_CONFIGURE_TIMING option (PHI Kanban #2719)
I added a new option <Project>_ENABLE_PACKAGE_CONFIGURE_TIMING to determine if
timing should be done for individual package configures.  This is defaulted to
FALSE in TriBITS and VERA.  This saves about 15s in the configure time for all
of VERA.

The code will also respond to <Package>_PACKAGE_CONFIGURE_TIMING to allow
timing indivdual pacakges.