Change final test regex to not break outer test regex
There was an evil bug in these unit tests.  The issue is that when I
added testing testing of the ctest properties, I added a check for
the final TRIBITS_ADD_ADVANCED_TEST() test criteria.  In doing so
I printed out "OVERALL FINAL RESULT: TEST PASSED".  The problem is that
the outer TAAT() command greps for this and if it finds it it says the
test is passing.  It turns out that I added a new test and changed the
number of passing tests and it just accepted it.

This commit changes the behavior of TAAT() so that in unit test mode it
sets the final regex as "OVERALL_FINAL_RESULT: TEST PASSED".  Now the outer
TAAT() test will only pass if the driver script prints that property.