Removing -VV from ctest because it disabled the test!
This was a similar case to my last commit.  You can't use -VV in the test
because if you one of the inner TRIBITS_ADD_ADVANCED_TEST() commands will
print "OVERALL FINAL RESULT: TEST PASSED".  If that happens, any of the inner
tests can fail and then the outer TAAT() will grep for that and show passing.

Warning, this change shows that there is a failing test!  My next commit will
fix that.