Adding ability for a TESTONLY libs to link to other TESTONLY lib (TriBITS #28)
I turns out that the old implementation never allowed you to link one TESTONLY
lib to another, at least not if it was in a different subdir.  You have to
bring in the include dirs, like TAE() does.

Also, I removed the duplicates from the TESTONLY include dirs which is
exported.

Finally, I cleaned up the logic that looks at correct usage of DEPLIBS.