Added c_util executable, make SpkgB required for SpkgC, fixed tests (TriBITS #28)
I added the the c_util executable in order to test creating an exec before any
libs were defined.

I made SpkgC have a required deps on SpkB because C.cpp assumed that SpkgB was
always enabled.  Not to worry, SpkgB has an optional library dependence to use
an as example and for testing.

I had to fix up some of the tests after these changes.