Adding back -VV to inner tests
Now that TRIBITS_ADD_ADVANCED_TEST() greps for a string that has the test name
in it, it is safe to add back the -VV option for the inner CTest runs.  Now if
some inner test fails, I can see the detail.

However, it appears this will not help for the ray tracer tests in the MixLang
package as they don't print anything, even if they fail.  I will have to do
something about this once I can confirm that these tests are failing on some
Trilinos test machines.