Removing deprecated EXPORT_LIBRARY_DEPENDENCIES() (TriBITS #35)
This was generating a file ${PROJECT_NAME}LibraryDepends.cmake that I don't
see documented anywhere as part of the export system.  My guess is that this
was an early version of an export system for CMake and is not being used
anymore.

If this does happen to break someone's project, they can just use the updated
export files.

Fixes #35.