Break up TRIBITS_GET_ENABLED_LIST() into for specific functions (TriBITS #38)
The problem with the old TRIBITS_GET_ENABLED_LIST() function is that it only
worked for explicit "ON" and "OFF".  That caused a problem when people used
TRUE and FALSE (see TriBITS #38).

Using these new functions is clearer and should also be faster because the if
statements in the inner loops are much simpler.

I have not added tests to prove that TRUE and FALSE work but all of the
existing tests all pass.