Replace APPEND_STRING_VAR(...) with LIST(APPEND ...) to write XML deps file (PHI Kanban #2718)
Here I have replaced the usage of the function APPEND_STRING_VAR() with
LIST(APPEND ...) which appears to be a lot faster.  The issue is that
APPEND_STRING_VAR() was creating new strings all of the time.

This reduced the cost of writing the XML dependencies file for Trilinos from
13 sec to under 6 sec.  Hopefully the impact for VERA will be even larger.

This was a super easy refactoring and there is a test case that protects this.