Adding infrastructure for automated testing of TribitsTplFindIncludeDirsAndLibraries.cmake (TriBITS #41)
I now have the infrastructure to be able to quickly test the full functioning
of TRIBITS_TPL_FIND_INCLUDE_DIRS_AND_LIBRARIES() in fast little unit tests.

I added a new mock TPL HeadersAndLibsTpl with empty header and library files
that the CMake commands FIND_PATH() and FIND_LIBRARY() seem to work wtih.
These commands are run as cmake -P scripts

This test shows that FIND_LIBRARY() does not look inside of the library files
at all, it just looks for the file names based on the prefix and suffix.

After this I will add a bunch of unit tests to assert the behavior of
TRIBITS_TPL_FIND_INCLUDE_DIRS_AND_LIBRARIES() for as many cases as I can think
of.