Fix nasty bug where parent shows as disabled with subpackage enabled
Here I fixed a nasty bug that was seen while trying to put together a tarball
for VERA.  The issue is that if a parent package is hard disbled but a
subpackage is enabled, it was not setting <Project>_ENABLE_<ParentPackage>=ON
at the end of the enable/disable logic.  This did not cause a problem for
regular configure processing because <ParentPackage> was being processed if
any of its subpackages were enabled.  However, the variable
<Project>_ENABLE_<ParentPackage> was still set to OFF.  This resulted in an
exclude added to the tarball so that the parent pakage directory was excluded
even though one of its subpackages was included!  This was super evil.

The fix was simple, you just force <Project>_ENABLE_<ParentPackage>=ON at the
end of the enable/disable logic if any of its subpackages is enabled even if
<Project>_ENABLE_<ParentPackage>=OFF.  This will avoid the exclude being
created for the directory of the parent package.

I added a test to demonstrate the desired behavior.  I worry that this is my
only test for this behavior and it is for the TribitsExampleProject.  I just
have to make sure that I preserve this test somehow.