Fixing return code for runSysCmndInterface()
This was using subprocess.Popen() incorrectly.  I found this out while trying
to code some other test code up.  After reading the documentation for
subprocess.Popen, I figured out how to get the return code correctly.

The reason that I never saw this problem before was that
runSysCmndInterface(..., rtnOputout=True, ...) was really only used in
CheckinTest_UnitTests.py where all of the commands where mock commands, not
real commands.  For that case it worked fine.

I had to fix a few of the tests for this change.  I can't understand why I did
not see this bad behavior in these tests.  The return code for a program can't
be None!