Split lines of shortlog -s on any whitespace (TriBITS #45)
So it turns out that 'git shortlog -s' uses a tab, not a space to separate the
number of commits from the author's name with '<num_commits>\t<author_name>',
not just a space.  This fixes that.

I updated one test to prove that both tabs and spaces work for the split.

Build/Test Cases Summary
Enabled Packages: TriBITS
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=177,notpassed=0 (0.80 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)