Process CallbackDefineProjectPackaging.cmake before CallbackDefineRepositoryPackaging.cmake
Here I moved the processing of the repos's
CallbackDefineRepositoryPackaging.cmake files near the very end right before
INCLUDE(CPack) is called.  This is so that the code in the
CallbackDefineRepositoryPackaging.cmake files has access to all of the
project's setting needed to copy files on the source distribution tree.

The driver for this is that MPACT needs to include a generated file in the
tarball but to do that, the project-level CPACK vars like
CPACK_SOURCE_GENERATOR and CPACK_PACKAGE_NAME must be fully defined.

Note that I did not change any tests as part of this change.  That is because
the TribitsExampleProject file CallbackDefineRepositoryPackaging.cmake does
not need these vars and no other part of the testing checks the order of the
processing of these files.

After this commit, I will look at adding code to
CallbackDefineRepositoryPackaging.cmake that requires these vars to be set and
then will copy a file into the tarball and verify that it comes back out of
the tarball as well.

Build/Test Cases Summary
Enabled Packages: TriBITS
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=177,notpassed=0 (0.80 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)