Don't look for diff in .gitigore when missing from orig source tree (PHI Kanban #3529)
When TriBITS itself is copied into the tarball, TriBITS has default ignores
for .gitignore files so the test that looks for:

  Only in .*/TribitsExampleProject: .gitignore

fails when it is run from the untarred source.

This was easy to fix by only looking for this diff when .gitignore exists in
the source tree.

I verified that this test when run from the orgininal git source tree looks
for this diff output string in the tarball testing.