Updating bulding developers guide (TriBITS #26)
This is just enough to get TribitsDevelopersGuide to build again.  Note that I
had to move a test file into this directory so as the documentation does not
depend on the tests dir.  I want to be able to build this documentation having
just the 'tribits/' directory and nothing else.

I also got the checkin-test.py script to run with --help again.  Just had to
fix a few paths.