Add Version.cmake file and get TribitsPackageFilePathUtils_UnitTests running (TriBITS #26)
Turns out that since the TriBITS project did not have a Version.cmake file,t
the variable ${PROJECT_NAME}_ENABLE_DEVELOPMENT_MODE_DEFAULT was not getting
set to TRUE and the variable ${PROJECT_NAME}_ENABLE_DEVELOPMENT_MODE was
getting set to FALSE.  This was excluding this one test.  It was not getting
run at all for a long time since I created a stand-alone TriBITS project
build.