Moving out non-core functionality with calls (TriBITS #26)
The generation of XML dependnecy files is not core functionality so I moved it
out into a separate module and call it if the directory tribits/ci_support/
exists.

The 'dashbaord' build target is not core functionality so I moved it out into
a separate module and call it if the directory tribits/ctest_driver/ exists.

The goal is the shrink the size of "core" TriBITS as much as possible.