Add temp symlinks to old TPL location (TriBITS #26)
I order to smooth the transition to the new TriBITS structure, the old
directory and find modules (symlinks) are restored.  This will allow me to
push the updated TriBITS and then we can work on getting the various
components to update their TPLsList.cmake files to point to the new locations.