Make repo update robust to forced branch updates
This changes to do the following for each git repo:

$ git clean -fdx
$ git fetch
$ git reset --hard <tracking_branch>

I tested this manually by running one of the VERA CTest driver scripts and
verified that it does indeed clean out the untracked and ignored files and it
resets the branch correctly.  This requires that each repo be on the correct
branch.  If it is not, then the script should return false.

I also upated the unit tests for this.  They are a little stronger but not
fantastically strong.