Don't print CATEGORIES trace unless tracing enabled (TriBITS #47)
I was missing a if() clause for <Project>_TRACE_ADD_TEST.  Therefore, it was
always printing when a test was not added because CATEGORIES did not match.

The unit tests did not need to be changed because the tests were only checking
if the message was printed when tracing was enabled.  I don't have a nice way
to check that such output does not exist.