Record deleted and change type as "modified" files, captitalize legend
I noticed that when I deleted some files or changed their type (e.g. from a
symlink to a real file), they did not show up under "M" for modified files.
Deleting a file and changing its type is sure like modifying it for the
purpose of this script.

I also added capitalization to some of the lines in the printed legend.  I was
inconsistent before.