Move TriBITS Version.cmake file to tribits/, include, and print version
This is the beginning of an attempt to provide TriBITS projects with a way to
inspect the TriBITS version.  I moved the TriBITS/Version.cmake file into
TriBITS/tribits/ and made a symlink back.  That way, the tribits/Version.cmake
file will get snapshotted or installed with the rest of the tribits/ directory.

What this is currently lacking is a way to track TriBITS versions precisely.
This will be figured out later.