Fix TribitsDumpDepsXmlScript.cmake for allowed missing packages
A while back I added support for directly listing possibily missing packages
in a PackagesList.cmake file and added good unit tests to show that it behaved
correctly for a *basic configure*.  However, I had never tested the behavior
of the checkin-test.py script on a project that used this feature (because it
was a new feature).  Turns out that the script TribitsDumpDepsXmlScript.cmake
that was called cased a failure when run in this mode.

The reason was that the script TribitsDumpDepsXmlScript.cmake was settting
${PROJECT_NAME}_IGNORE_PACKAGE_EXISTS_CHECK=TRUE which was resulting in the
missing package being added to the list of anyway which then resulted in an
error when attempting to read in the Dependencies.cmake file.  I have no
memory of why this variable was set in this script.

To add a failing test, I added the package AlwaysMissing to the
MockTrilinos/PackagesList.cmake file.  This resulted in lots of tests failing.

To fix the the failing tests, I just removed

   SET(${PROJECT_NAME}_IGNORE_PACKAGE_EXISTS_CHECK TRUE)

from TribitsDumpDepsXmlScript.cmake and TribitsCTestDriverCore.cmake scripts
and then all the tests pass again.

It makes no sense in any real usage of this code to ever add an allowed
missing package to the list of defined packages.  The only case where it makes
sense to set ${PROJECT_NAME}_IGNORE_PACKAGE_EXISTS_CHECK=TRUE is in a few unit
tests that *don't* read in the Dependencies.cmake files.  There are a few unit
tests that do this so you will see that variable set in those tests.

To see the failing tests, just revert the change in this commit to the files
TribitsDumpDepsXmlScript.cmake and/or TribitsCTestDriverCore.cmake and watch
the tests fail.

Build/Test Cases Summary
Enabled Packages: TriBITS, CTeuchos
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=186,notpassed=0 (3.21 min)
1) SERIAL_RELEASE => passed: passed=181,notpassed=0 (1.10 min)
Other local commits for this build/test group: 6f999fd