Fix failed TPL find can be found on reconfigure  (TriBITS #41)
With my previous refactoring, if it failed to find the include dirs or the
libs, it would fail the reconfigure even when it could find the right dirs and
libs.  This was because the configure that failed set cache vars

  TPL_${TPL_NAME}_INCLUDE_DIRS=""
  TPL_${TPL_NAME}_LIBRARIES=""
  TPL_${TPL_NAME}_NOT_FOUND="TRUE"

and they never got changed.

Fixing this was trivial.  You just add FORCE to the set of the vars:

  TPL_${TPL_NAME}_INCLUDE_DIRS
  TPL_${TPL_NAME}_LIBRARIES

when a find is done.

The ADVANCED_SET( ... FORCE ) is never called if these vars are non-empty.

The hard part was getting auomated tests in place that show this problem.  For
this, I had to make the ProcessEnabledTpls tests do a cmake configure with a
cache file so that we can see the impact of the cache vars on reconfigures.