Try to fix generation of build ref doc
For some reason, the bash function generate_git_version_file was failing and
aborting the call to 'make'.  This was aborting the update of the build
reference guide.

Since we don't rebuild the guides just based on an updated git version, we
don't need to create a temp file so there is no reason to even create it.
Just create the final TribitsGitVersion.txt file.