Fix env vars handling for running processes (PHI Kanban #3262 and #3217)
Turns out that if you set 'envirornment' to the runSysCmnd() that would
replace the entire environment!  That is not what I intended!  Therefore, I
changed this argument name to 'extraEnv' and I get the full env then update it
with extraEnv.

There was only one place in TriBITS that I could find that was using this and
that was tdd_driver.py.  I fixed the usage.