Updated to install gcc-4.8.3 (PHI Kanban #3262 and #3217)
I made a lot of changes to this script.  It now can only install gcc-4.8.3.
It will no longer install gcc-4.6.1 or gcc-4.5.1.

The source for gcc-4.8.3 is now stored in a single git repo untarred.  This
includes updated gmp, mpc, and mpfr (see the source).  There is no reason to
untar anything.  This was needed because the full tarball was over 100M and
github will not allow that.  Also this allows you to version control changes
(like upgrading gmp, mpc, and mpfr).

I also made it so that if you use --checkout with an existing local dir, it
removes it and does the git clone again.

Also updated to use the new --parallel argument to set -jN for make.