Merge commit 658810931b1440173877b3e85e482d5998e88c73 from bartlettroscoe
Add failing test demonstrating broken <Package>Config.cmake in build tree (TriBITS #46)

This test shows that the files <Package>Config.cmake generated in the build
tree and meant to be used by clients to get access to software in the build
tree (not the install tree) are broken.  It can't even find the subpackage
include files.