Remove quotes around two variables (CMake 3.1)
As of CMake 3.1, CMake defines a new policy CMP0054 that warns that quoted
strings like "SOME_NAME" are first looked up as a variable and then the value
of the variable is substuted.  That is very bad (in my opinion).  There were
just two places where I was doing this but that was not necssary or desired.
Therefore, I removed the quotes.